Restructure events directory & move tests
