+++ connectivity/drivers/emac/TARGET_ARM_FM/COMPONENT_LAN91C111/fvp_emac.cpp
@@ -24,7 +24,7 @@
 #include "mbed_interface.h"
 #include "mbed_assert.h"
 #include "netsocket/nsapi_types.h"
-#include "mbed_shared_queues.h"
+#include "events/mbed_shared_queues.h"
 
 using namespace std::chrono;
