+++ connectivity/drivers/emac/TARGET_NUVOTON_EMAC/numaker_emac.cpp
@@ -28,7 +28,7 @@
 #include "mbed_interface.h"
 #include "mbed_assert.h"
 #include "netsocket/nsapi_types.h"
-#include "mbed_shared_queues.h"
+#include "events/mbed_shared_queues.h"
 
 //#define NU_TRACE
