+++ connectivity/drivers/emac/TARGET_NXP_EMAC/TARGET_IMX/imx_emac.cpp
@@ -40,7 +40,7 @@
 #include "mbed_interface.h"
 #include "mbed_assert.h"
 #include "netsocket/nsapi_types.h"
-#include "mbed_shared_queues.h"
+#include "events/mbed_shared_queues.h"
 
 #include "fsl_phy.h"
