+++ connectivity/drivers/emac/TARGET_NXP_EMAC/TARGET_MCU_LPC546XX/lpc546xx_emac.cpp
@@ -24,7 +24,7 @@
 #include "mbed_interface.h"
 #include "mbed_assert.h"
 #include "netsocket/nsapi_types.h"
-#include "mbed_shared_queues.h"
+#include "events/mbed_shared_queues.h"
 
 #include "fsl_phy.h"
