+++ connectivity/drivers/emac/TARGET_STM/stm32xx_emac.cpp
@@ -24,7 +24,7 @@
 #include "mbed.h"
 #include "mbed_interface.h"
 #include "mbed_assert.h"
-#include "mbed_shared_queues.h"
+#include "events/mbed_shared_queues.h"
 #include "netsocket/nsapi_types.h"
 
 #include "stm32xx_emac_config.h"
