+++ connectivity/libraries/ppp/source/ppp_service_if.cpp
@@ -20,7 +20,7 @@
 #include "mbed_interface.h"
 #include "mbed_trace.h"
 #include "netsocket/nsapi_types.h"
-#include "mbed_shared_queues.h"
+#include "events/mbed_shared_queues.h"
 #include "mbed_poll.h"
 #include "ip4string.h"
 #include "ip6string.h"
