+++ connectivity/netsocket/source/nsapi_dns.cpp
@@ -24,7 +24,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <stdint.h>
-#include "mbed_shared_queues.h"
+#include "events/mbed_shared_queues.h"
 #include "events/EventQueue.h"
 #include "netsocket/OnboardNetworkStack.h"
 #include "rtos/Kernel.h"
