--- events/source/.mbedignore
@@ -1,1 +1,0 @@
-tests/*
