+++ events/tests/.mbedignore
@@ -1,0 +1,2 @@
+UNITTESTS/*
+unit/
